dGrid: Added missing sort() hook method definition. Previously this was defined only in the datanav grid, even though it was called from this class.
dPemMixin: Fixed a potential divide by zero error when doing gradient drawing on extremely small areas.

dPageFrameMixin: fixed a layout bug that prevented calls to layout() from being passed through to the pages.

dEditor: improved the order of the font attribute initialization so that it would be possible to pass font properties in the constructor call.

dListControl: prettied up some of the Dynamic* calls.