Changed dBizobj's __setattr__ to not try to call setFieldVal() unless we are out of the init phase. setFieldVal() was being called unnecessarily just to fail, raise an exception, and get caught by __setattr__.  Removed a bare try/except block from dBizobj.__setattr__ which was hiding any attribute related exception from being properly progagated. Instead, we rely only on the return value from dBizobj.setFieldVal. Changed dBizobj.setFieldVal to intercept only the dBizobj.NoRecordsException instead of catching and hiding all exceptions. Fixed dCursor.setFieldVal to raise NoRecordsException instead of dException in this case. The result is that unexpected exceptions will no longer be hidden.
In addition to the above noted changes to dBizobj.setFieldVal, it will now only
try to call dCursor.setFieldVal() if the field is part of the cursor's structure.

Moved some attribute defines from dBizobj._beforeInit to the top of the init
cycle. Removed the setting of dBizobj._autoCommit as this isn't used anywhere
(dBizobj.AutoCommit is just a proxy to dCursor.AutoCommit).

Added codes "G" for long and "F" for float to the daboTypes. Added a mapping
from the pythonType of str back to the daboType of "C" (which in turn maps to
unicode still).

Changed the relevant code in dbMySQL to get floats and longs as appropriate 
instead of decimals and ints as it was doing before. Shied away from making
any changes to dbFirebird and dbPostgres as I have no experience with those
databases.

Refactored the code in dCursor.execute() that coerces the fields into the
proper types (e.g. converting str to unicode) into a new function 
correctFieldType(). Refactored dCursor.storeFieldTypes() somewhat. Set 
setFieldVal() to ignore reporting data type mismatch for fields that are in
NonUpdateFields, which are likely calculated and so it doesn't really matter. 

Added new FieldNotFoundException, and set setFieldVal() to raise that in that 
event.

Cache the structure description so it isn't recursively getting it for every
record during the storeFieldTypes() call. We only need to get it once, after
each requery.