Reordered the superclass definitions for all classes that descend from a wx class. This appears to solve the property clobbering that was happening starting with wxPython 2.7.
Dabo 0.7 will be required for wxPython 2.7 and beyond, because I don't 
believe we should backport this to the stable Dabo (in case it causes 
other trouble not yet discovered).

wxEvt.KeyCode is now a property instead of a method, so I fixed the 
one place I found that used it.

Fixed a spelling error I found in dDialog.LastPositionInSizer property.

wxPython 2.7 no longer has a SetTitle() method for some classes where
it had it before, so I wrapped some code in a try block to compensate.

ImageMixin previously wasn't a new-style class. Fixed.

With these changes, I can get most stuff starting and running with 
wxPython 2.7 (and still wxPython 2.6). However, I've only tested so
far on Ubuntu Edgy.