Fixed a bug in the Selection property of dTreeView. If MultipleSelect is False, it should return a single node, not a 1-element list.
Added the 'ShowBox' property to dRadioList. It controls whether there is a visible box with optional caption displayed around the buttons. 

Fixed a bug in dEditor that was throwing errors when the lexer was being set before the Parent property was set.

Added methods to dSizer and dBorderSizer that will return the corresponding bordered/non-borderd class. This was necessary for adding the ShowBox property to dRadioList, since I needed to be able to replace a border sizer with a non-border sizer, and vice-versa. I couldn't hard-code the classes, since in the Class Designer they need to be designer-specific subclasses.