Added the method 'showContainingPage()' to dPemMixin. Calling this on any object will cause any pageframe-type controls inside of which the object is contained to switch their selected page to the one containing the object. This is recursive, so that even in pageframe-within-pageframe scenarios, the object will end up being visible.
Added auto-sizing of columns to dListControl. Added get/setItemForeColor() methods; it already had the get/setItemBackColor() methods.

Added the method 'showCurrentLine()' to dEditor. This will scroll the editor so that the specified line is visible.

Fixed a minor bug in dRadioList that happened if you instantiated it with ShowBox=False passed to the constructor.