Backported to stable:
Added dabo.ui.getImagePath(), which will return the full path of an image given just its name. It will look in the standard Dabo directories, the app HomeDirectory, the current directory, and all of the sys.path directories. It can optionally return the path in URL format.

Updated dHtmlBox to use the getImagePath() method in order to resolve <img> tags. Now you can just specify the image name in the "src" attributer, and if it is in one of the standard paths noted above, it will be converted to a full file URL.