Fixed the 'Replace' function in the Find/Replace dialog when working with dTextBox/ dEditBox controls.
Removed the restrictions on incremental search being only for ASCII characters. I have tested this with some non-US keyboard layouts and have not gotten any errors, but I don't have any Cyrillic or other non-ASCII data sources to test against. If you have such data, please try incremental search with it.

Changed the way that the incremental search status text was displayed. Previously, the act of moving to the selected record wiped out the status text and replaced it with the 'Record n/x' message. 

Changed the default search delay from 300 ms to 500 ms. I found that 300 ms was fairly quick, and frequently got the wrong record because a slight extra delay between keystrokes reset the time. A couple of people at some recent demos also had the same experience, so I've changed it to a value that seems to work better for many others.

Added a small border to the Requery button on the datanav Select page. The button was appearing clipped on OS X.

Changed the comments in dSpinner so that the Children docstring matches that of dPemMixin, with the explanation of the code moved to the _getChildren() method.