Removed the MultipleSelect property from the initProperties for dTreeView. It isn't needed there, and in fact prevents it from working correctly.
Changed the code in the size properties in dPemMixin so that if the object is a panel, the SetMinSize call resets the minimum size for the panel. Previously, this was causing panels to be able to be set larger in sizers, but not to shrink back down as needed.

Added a 'resetMin' parameter to the layout() method of dPanel. This is related to the issue mentioned above; together they allow dynamic sizing of panels to both grow and shrink the panel.

Also added a performance improvement to only check for the platform once in the life of a panel instead of every time it was sized.

Added setting of the height and width attributes for drawn bitmap objects, so that code can determine their sizes.

The append/insert methods for dMenu have a 'bmp' parameter. I added a 'picture' parameter, which is more consistent with the rest of dabo.ui in treating image values. 'bmp' still works, but unless you are passing an actual bitmap instead of a path to an image, 'picture' is the correct parameter to use.