Added this newly-changed-licensed tool to dabo.lib. Thanks to Marius Gedminas for this excellent program.
To use it to track down slow parts of your code, add the following import statement:

from dabo.lib.profilehooks import profile

Now let's say that the function is the bar() method of your Foo class. All you need to do is decorate bar() as follows:

		from dabo.lib.profilehooks import profile
		class Foo(object):
			@profile
			def bar(self):
				<bar code>

I haven't played with the coverage part of this yet, but I would assume that it would be equally useful.