Completed the basics for the createMenuBar() function. You can now pass in a form reference, so that bound functions can be written in the format "form.someFunc", and it will correctly resolve that to the actual form method. You can also bind to application methods using the 'app.method' syntax, createMenuBar() will automatically determine the app reference.
There is an additional parameter named 'previewFunc', which accepts a callback function that will be called instead of the normal bound function. This is for testing during design time, where you might want to test that the menu is configured correctly, but without having the actual functions enabled, such as if you had a 'Quit' menu item.

Modified the drawRectangle() method so that if you pass -1 for the width or the height, it will use the full width/height of the parent object.