Fixed some issues with using the dImage control for handling blob data.
Improved the performance of dImage by removing some repeated calls to methods such as _showPic() that were being called several dozen times per change to the image.

Fixed the attProperties parameter issue in dColumn.

Changed the form's requery to not use the progress dialog, and just call the bizobj requery directly. I added a dabo.ui.busyInfo() call in case of long requeries, which was the original intent of the progress dialog.

Added a layout() method to dStatusBar. It can contain sizers and other controls, so this makes it consistent with other Dabo containers.