Added the ability to pass parameters to dabo.ui.createForm(). You can now specify runtime properties in addition to those defined in the cdxml file.
Changed the logic for applying attProperties. Previously, anything in attProperties overrode any matching key in the properties dict. E.g., if a form had a caption of 'Foo' in the cdxml, and you then specified dabo.ui.createForm("whatever.cdxml", Caption="Bar") on creation, the value in the cdxml would 'win', and the form would have the caption of 'Foo'. With this change, the explicit property passed to create form will prevail.

These correction address the Tracker Issue #0273, in which MDI children couldn't be created, since you couldn't pass the parent reference.