Added the ability to pass event bindings to the constructor of an object.
They must be in the form of 'On<name of Event>=method'; e.g.:

dabo.ui.dButton(self, Caption="Close", OnHit=self.onClose)

I think that this is very much in the spirit of auto-binding of events, and a lot cleaner-looking and easier to understand than the on<evt>_<RegID> construct for contained objects. It's also consistent with the passing of 'bindfunc' to menu items' constructors. Finally, I have a ton of places where I create a button and then do a bindEvent() to the Hit event immediately afterward. This eliminates that extra code.