OK, this is another major change that ties into yesterday's change that allows event binding to be passed in the constructor call for an object.
dMenu was written with a 'bindfunc' parameter to its menuitem-adding methods that specified the method to call when that item was selected. It has now been deprecated; instead, pass an OnHit parameter. 'bindfunc' will still work, but will throw a deprecation warning. It will be removed by 0.8.

All framework files that used bindfunc have now been updated to use OnHit.