Added the method 'getOriginalImgSize()' to dImage. Since the control scales the image, it is handy to be able to retrieve its normal size.
Removed the restriction in dImage that standard graphic references can't be used. I don't know why that was put there, and can't see any benefit.

Modified the form mixin class to add a 'MenuBarFile' property. If this is an mnxml file, it will use this for its MenuBarClass.

Improved the creation of menus in dabo.ui.createMenuBar(). The Picture property was being ignored, and I changed the 'Function' property to 'Action' instead, as this just seems simpler and clearer. Also cleaned up an old 'self.' reference in the dabo.ui code.