This is a major rework of the mementos system in dCursorMixin. Previously, we copied a memento for every record at the time of requery, and when it came to check to see if there were any changes to any records, we crawled through each parent record, then through children, their children, etc. etc. testing each memento against each field. Now, we only record a memento when the field value changes.
This commit also includes unit tests for dCursorMixin and dBizobj that test the
memento system.

Added scanRows() and scanChangedRows() functions to dBizobj, which do the same
as scan() but only for the applicable rows.

Added an _index() function to dDataSet, that returns the index (row number) of
the passed record object.

Reworked how dBizobj/dCursorMixin.DataStructure gets the field type information.
Specifically, reworked dBackend.getStructureDescription() to first look at the
cursor.description, and then use getFields() to fill in any missing information.

Removed a duplicate conversion to dDataSet in dCursorMixin. Tried to find and
fix instances where we raise dException.dException and we should be raising 
more explicit exceptions such as dException.FieldNotFoundException.

Please note that I tested against MySQL and sqlite, not against Firebird, 
PostgreSQL, or MS-SQL.