Because my edits of the datanav framework cause all kinds of backwards- incompatible changes, I've decided not to mangle the current datanav but instead to create a new datanav2 library.
Added datanav2, which is copied from my branch, and marked the original
datanav as deprecated. Old apps need to nothing to run unchanged against
the old datanav. New apps will be generated by the AppWizard to run 
against datanav2.

A brief summary of changes:

+ Removed everything having to do with FieldSpecs and RelationSpecs.
+ Removed Form.creation(), beforeCreation(), and afterCreation().
+ Removed Form.getEditClassForField().
+ Removed Form.getSelectControlClassForField().
+ Removed Form.getSelectOptionsForField().
+ Simplified much of the underlying code.
+ Added boilerplate test code to the form classes and page classes
  to assist with modifying the layout and being able to quickly see
  how it looks when run.

The general theme being that the generated code is meant to be edited by
the developer to achieve the desired results. Eventually, we'll hopefully
get to the point where there is no datanav framework, and the generated
code directly subclasses from dabo.ui.