OK, I've been working on this on and off for the last week or so, but finally have something that seems to work OK for now.
This is the revised dDateTextBox that removes all of the formatting options that was causing a lot of problems. I thought that without our manual formatting that the date would display in the current locale, but apparently it displays in the standard datetime.date format of YYYY-MM-DD. You can enter a date in a standard format like MM/DD/YYYY or YYYY.MM.DD, but it will automatically reformat to YYYY-MM-DD.

I've tested this a lot, and it seems to work much better, but please test this in any of your apps that use dDateTextBox. Feedback is welcome, of course.