Added VirtualFields functionality to dBizobj and dCursor. VirtualFields don't really exist in the dataset: they map to functions that get called during getFieldVal() to determine the value of the virtual field.
Removed cursor.setSQL() and all calls to it, as the code inside it had been
commented out for quite a long while now. 

Added support for automatic conversion of dates and datetimes saved as strings 
in the backend to datetime.date and datetime.datetime objects in the cursor.
This comes in handy with sqlite, where I have date fields stored in YYYY-MM-DD
string format.

Refactored getDataSet() to only have to iterate the records once, and to include
any virtual fields in the result.