This is a result of a request made by John F for dTextBox objects to have the ability to limit the amount of characters that can be entered.  I added a TextLength property that will dynamically adjust the value in a way similar to the ForceCase Property.
I created a full suite of unit tests for this property.  The only thing the UT doesn't do is check for the dynamic updating when a user enters characters in a text box.  I checked that by hand and it seems to work.  Also, I checked that it was playing well with ForceCase and everything seemed to be working well.  However, I would encourage people to play with this and report any bugs like usual.

This only affects string textboxes.  I figured it would be useful so you don't pass a bizobj a 10 char string when the db field is 2 characters long.  Let me know what you think.