Fixed setHeight and setWidth to not change the opposite dimension to a fixed value. Instead, -1 is passed which tells wxPython to not change it from the current value, but also that wxPython can still automatically adjust it when necessary, such as when we've zoomed in on the font.
Previously, setting a fixed width on a textbox also would set a fixed height,
and when we zoomed in, the textbox wouldn't grow in height to accommodate the
increased font size.

Originally, I thought the growing behavior was dependent on having a sizer
that could grow, but today I discovered it was dependent on not having a 
fixed size set on the object.