Followed through with the saving-blank-new-record thing. I didn't take into account the filling of FK fields or anything the user might have put in the onNew() hook (according to Ed, anything done there shouldn't be considered a change).
I also didn't take into account the diffs made when saving a row. For new
records, we need a diff that includes all records, not just the ones with
mementos, because the backend needs to know about the new fields.

I think this problem is licked now, but let me know!