Changed the logic surrounding field-level validation to distinguish when a rule has passed versus the case where there is simply no rule. This is implemented with the new BusinessRulePassed exception class, which is caught by dForm, and then calls onFieldValidationPassed(). By default, this method will clear the status text set by a prior onFieldValidationFailed() call.
Also updated the data controls to validate changes if the user changes the field value back to its original value, so that we can determine that the 'new' value will also pass validation, and clear the StatusText.

Changed the alias used for dConstants from 'k' to 'kons', as I feel it is more readable.