Dynamic props on dColumn were not working: fixed.
Began work getting cell properties set up. IOW, being able to set the
ForeColor of a specific row in a column (a single cell), including
dynamic cell properties. Currently, it works, but the only property
I've added yet is CellForeColor and DynamicCellForeColor. We'll need
to go through and add all the apropos properties such as alignment, 
font, and color properties but I wanted to test this before going
much further with it.

Obviously, performance has the potential to degrade, depending on
what the developer codes in their dynamic function handler, as 
dGrid will ask for the dynamic property value for every displayed
cell. But, it does behave virtually, much the same as getting the
values to display in the cells, so for simple things it should be
okay. 

For the Cell Dynamic Props, the handler needs to accept a 'row'
argument (dGrid won't move to the row automatically, which is a 
good thing, because you don't want all kinds of automatic child
requeries taking place if your function only wants to alternate
row colors).

For simple coloring based on field values in the database, the
best approach would be to call biz.getFieldVal(fld, row) instead
of actually moving to the row if you can avoid it.