The problem with the dForm test was that we now cache calls to update(), to avoid repeated unnecessary updates. So if more than one update() is called within 100 ms, only one call will get processed, in the next 100 ms interval. This is great for an actual UI, but for the test script, which doesn't wait for 100ms before running its tests, it isn't good.
I added a force argument to both dForm.update() and dForm.refresh();
when True, the call happens immediately like before. I can see this being
needed from user code as well in some cases, if the code relies on the
form being updated after a coded operation. I also fixed the docstrings
for those methods to actually explain what they do, instead of commenting
on the need to cache the calls.

Fixed the test to explicitly call update(True) in the needed places.