Modified the Log class to contain levels (Database Activity, Debug, Info, Error, Exception, Critical), as well as be accessed directly by dabo.logInfo("msg") or dabo.logError("msg") or dabo.logDBActivity("msg").
All of the dabo.logXXX methods also accept a second bool argument, which if True, will output the message to the console reguardless of the logLevel

Added two new methods addLogLevel and delLogLevel, which accept a single argument, and will add or remove that log type from the log level.

new method addCustomType will allow people to add their own log types to the logging system.