After this morning's reverts, I was still finding at least one error (the metaclass-base one that I discuss in the tracker item).
Because of all the noise in the 3194 commit, due I guess to Dj's editor fixing whitespace at
the end of lines, I gave up trying to locate the source of the problem in the diff.

This commit reverts trunk back to where it was at 3193, but includes Ed's 3195 commit
on dSizerMixin.