Added the ExpandToFit property. When True (default), the column specified in the ExpandColumn property will be resized to fit the width of the control.
Added the ExpandColumn property. This is either a numeric value indicating a column position, or 'LAST' (default), for which the rightmost column will be used. Only has an effect if ExpandToFit is True.

Added the SortColumn property. This controls which column is sorted when the control's sort() method is called. Clicking a column header changes this property's value to the clicked column.

Added the 'getColumnWidth()' method that simply wraps the wx version.

Added event handling for when the user manually resizes a column. This will raise a dEvents.ListColumnResize event.