This commit adds auto-update functionality to the Dabo framework. I've added some web services to dabodev.com that will report on the current status of the framework, and allow updates to be retrieved.
While the user interface for this is not set yet, a Dabo developer will be able to download the framework once without Subversion or any other software, and then control how often the framework will check the website for updates. If updates are available, the user will be given the option of downloading them and re-launching their app, or ignoring them for the time being.

Still lots of stuff to add, but this is working pretty well from within my domain. Next we need others to test it.