I have created the dLinePlot object.  This object uses the wx.lib.plot objects to create a widget that can plot line and scatter plots.  The widget has many features.  You can set the axis label, graph title, legend, trace properties, grid, and more.  It includes a test that you can run by running dLinePlot.py.  Not everything is implemented yet, but this is a nice basic implementation.
I added the dLinePlot import to the __init__.py.  It is in a try/except clause because when you import it in wx you need NumPy (Since the wx widget needs NumPy, I left the numpy import statement at the top).  This way it doesn't break existing functionality.

Please try and see what you think.  I've tried to keep the naming conventions and syntax akin to Dabo, so it should be straight forward.