Retranslated the rest of the Spanish strings that were, uh, let's say... not accurate before. :)  Not that I'm a native Spanish speaker, mind you, so if you know better please help us out.
Also, I figured out how to handle special unicode characters in the .po file
so that the accented characters make it to the user interface: write the 
encoded unicode byte sequences into the .po.

If you want to test this out the accented characters are in the status bar 
text of a couple items in the Edit menu.