Fixed a major security hole in dEditor.py, in which any code in the editor was executed every time an object reference was needed.
Fixed a glitch in dSplitter in which sash percentages could be set to zero due to integer division.

Added the code to restore the __builtin__._ function after lines were processed in dShell, but commented it out as it is not needed with the explicit 'from dLocalize import _' approach to localization.

Improved the event handling in dFoldPanelBar.