Added Larry's patch to dGrid, which should help keep grid column sizing from triggering processSort(), although I think this just hides a deeper problem.
Because the column hasn't been sized yet, getColByX() still has the old 
column sizes. However, we are getting x from the mousemove event, which will
have the position of the mouse at the point of the move. So, move the mouse
more than the fudge amount before mousemove fires, and our code won't realize
that the grid column is sizing.

I don't really know what else to do, save ask wxPython to provide an event
to let us know when a size action is happening instead of merely when it is
complete.