Refined the way that the Preference Dialog works. The major change was better integration of the Preference Manager code that already handles caching and persisting/discarding changes. If your preference page controls settings that are all stored under a single preference key, you add that to the PreferenceDialog's 'preferenceKeys' list, and AutoPersist will automatically be turned off for that key for the duration of the dialog. If the user clicks OK, the changes are persisted; if they cancel, they are discarded. AutoPersist is then returned to its original value.
Changed some names of handlers in the PreferenceDialog class to avoid conflicts with other existing methods of dialogs.

Added an onCancelPref() handler that is called when the user cancels, just as the onAcceptPref() handler is called when the user clicks OK.

Updated the web update page to use the preference keys instead of coding additional methods to do the same thing.