Modified the RowNumChanged event to also pass along the bizobj whose row number has changed; this way only grids tied to that bizobj will update.
Changed the DataSource property of the grid so that if it is a string, the first time it asks the form to resolve it, it replaces the string with the bizobj reference, eliminating repeated calls in the future.

Updated the datanav code that relied on the grid's DataSource being a string.