Fixed dHtmlBox to be able to display inline images on Windows, which was previously not working because apparently it can't handle drive letters in the path.
Note that this fix assumes that the image will be on the same volume 
(drive letter) as the one where the script is running.

This fixes DaboDemo on Windows which now properly displays the dabo icon. 
I could have sworn there was a ticket for this but now I can't find it.