OK, I believe that this cleans up a hidden bug that has been in the code since, well, forever.
When we set properties on UI objects, we make sure that they have been constructed first; if not, we hold those settings in self._properties, and then attempt to set them just before _afterInit() is called. For simple objects, this is sufficient.

If, however, we want to create composite objects, such as what I've been working on with a spinner that uses two Dabo objects, the call to _setProperties() is still too early, since the child objects may not have been created, and the properties involve calls to those objects. The problem I found was that the first line of dObject's __init__() wiped out the self._properties dict, so it was no longer possible to preserve those values for setting later on.

I changed the code in dObject to only set self._properties = {} if that attribute did not yet exist, but this caused all sorts of secondary effects in which property settings were called multiple times. This was typically not bad, except in the case of the RegID property, which cannot be set more than once per object.

The solution I came up with was to add a line that clears self._properties in the _setProperties() method. If controls are not yet constructed, their prop values will be re-stored in self._properties, and can be then set whenever it is appropriate.