Revision [3732] introduced a bug when forms with self.Centered=True would not have the names 'left' and 'top' defined. This corrects that.
Added some intelligent path resolution for connection file resolution. Connections are always stored in .cdxml files relative to the .cdxml file location, but if an app uses a .cdxml with the current directory different than the .cdxml location, it couldn't find the .cnxml. This change uses some relative pathing code to properly locate the .cnxml file.