Since [3832], the flushValue() system for textboxes has been doing things in the wrong sequence, resulting in changes not getting properly flushed to the underlying bizobj. I fixed this by reintroducing the _inFlush attribute and removing the callAfterInterval().
However, after doing this the dEditBox test started failing again on setting 
the Value to None. This turned out to be because dEditBox was inheriting from
dTextBoxMixinBase, which doesn't have the logic to deal with anything other 
than string types. I made dEditBox inherit instead from dTextBoxMixin, which
solves the problem.

It may we weird for someone to define an editbox to edit date values, for 
instance, but I don't think it is wrong to allow it to work, since
fundamentally the only real difference between dTextBox and dEditBox is 
that dEditBox can have multiple lines.