If a data control has its DataField or DataSource properties changed, clear the _oldVal flag so that an immediate subsequent call to flushValue() will work.
I have a textbox on one of my screens that gets bound to different fields
depending on other factors, including the user choosing an option from a
dropdown list. Prior to this change, the code to get the value to 
propagate to the bizobj, *and* to not change the value of the textbox
when the data field changed, was ugly. 

TODO: Add a force option to flushValue(), which defaults to True, in which 
case the value will get set in the bizobj regardless of whether the ui
thinks the value has changed or not. We'll need to change all framework
calls to flushValue() to add force=False.

This is so that when an appdev writes flushValue() explicitly, the flushing
of the value will unconditionally happen, which is likely the expected
behavior.