Fixed dabo.ui.callAfterInterval() to work properly with closures. The way callAfterInterval() is supposed to work is that if the function is called repeatedly, it'll only get called once, at <interval> seconds past the last time the function was called. However, with closures, it was getting called every single time (a wx.FutureCall was being added for every call, and prior ones weren't being stopped).
The problem was that each closure is a different function instance. Fixing
it to key on func_code instead solves the problem.

This doesn't seem to break anything, but be on alert for unintended 
side-effects.

Here's a contrived example of use common to the way I code where I was 
getting bitten by the problem:

#--
class PanTest(dabo.ui.dPanel):
	...
	def onHit_txtTest(self, evt):
		...
		self.update_something_else()

	def update_something_else(self):
		...
		def do():
			...intense stuff - updating multiple controls...
		...
		dabo.ui.callAfterInterval(100, do)
#--

If the user was typing fast, that onHit was calling update_something_else
many times per second, and because the do() function gets instantiated 
anew every time update_something_else() is called, we get multiple
FutureCalls instead of the desired single FutureCall.