Fixed the error reporting code to handle exception types that can have or lack the various attributes that provide information about the exception.
Added code to read in connection information residing in the directories around a cdxml files when it is opened in the Class Designer. This will help enable you to run the Class Designer from one directory, and open designs in a completely different location without having to hard-code the full path to the .cnxml file in your design.

Thanks to Roger L for catching both of these!