Found and fixed the problem I reported earlier today. I was creating new records in child bizobjs, and then canceling the parent. But the iteration of the changed rows was happening ascending instead of descending, so if the changed rows were [0,1,2,3], once rows 0 and 1 were deleted, rows 2 and 3 were now rows 0 and 1 and the 'invalid row number' exception would happen.
This obviously fixes some subtle problems that have been around for a 
while. I previously fixed this in dCursorMixin's version of getChangedRows(),
but that version doesn't get used in the case of child bizobjs being
present.

The really insidious thing about bugs like this is that they don't 
result in unhandled exceptions: the user gets a message like:

+-------------------------------+
| Cancel not allowed.           |
+-------------------------------+
| Invalid row specified.        |
+-------------------------------+

...which means my customers have been dealing with this for longer than
I've known about it, since they are the ones using the app for real.
Had an unhandled exception been raised, my global error handler would
have emailed the traceback and other useful information.

I don't have a suggestion on how to be able to better debug the db and
biz layer at this point; this is just food for thought.