Added a ToolTipText property for Nodes. This way we can set the tooltips on a per-node basis.
Added the UseNodeToolTips property on the tree. When this is True, the ToolTipText is set to the node currently under the mouse; if there is no node, the tree's ToolTipText is displayed. If the node does not have a specific ToolTipText set, the node's Caption will be displayed in the ToolTipText.

Added the ImageSize property to the tree. This is a two-tuple of ints, with a default of (16, 16). This will ensure that all images added to a tree are consistent in size; some of the standard bitmaps are 16x15 or otherwise not acceptable for use in a tree, especially under Windows.

Added the option 'clearImageList' to the clear() method of the tree.

Greatly improved the makeDirTree() method. Added two parameters: 'ignore' and 'expand'; the first allows you to specify file patterns to exclude from the tree (e.g., "*pyc" to exclude compiled scripts); the latter will expand the tree once it is built. Also added default folder/file images to the generated tree. The nodes also include the full path to the file in their ToolTipText.